feat: Enhance ML capabilities and analytics
- Updated project files to include Microsoft ML and Polly packages.
- Extended `IMLAnalyticsRepository` with methods for interaction counts
  and user engagement scores.
- Changed `Interests` property type in `UserInterestDocument` to
  improve precision.
- Implemented new analytics methods in `MLAnalyticsRepository` using
  MongoDB's aggregation framework.
- Modified `MLContentFeatureRepository` to use `double` for
  `qualityScore`.
- Added new classes/interfaces for machine learning features,
  including `PostSummaryResponseML` and `IContextAwareFeedService`.
- Implemented `ContextAwareFeedService` for personalized feeds
  based on user context.
- Introduced services for feature calculation, model training, and
  monitoring.
- Updated `FeedController` with endpoints for smart, nearby,
  trending feeds, and similar posts.
- Overall structure enhancements for advanced analytics and
  personalized content delivery.

+++ Camply.Application/MachineLearning/Interfaces/IMLContentFeatureRepository.cs
@@ -7,7 +7,7 @@ namespace Camply.Application.MachineLearning.Interfaces
     {
         Task<MLContentFeature> GetLatestContentFeatureAsync(Guid contentId, string contentType, string featureCategory);
         Task<List<MLContentFeature>> GetContentFeaturesAsync(Guid contentId, string contentType);
-        Task<bool> UpdateContentFeatureAsync(Guid contentId, string contentType, string featureCategory, string featureVector, float qualityScore);
+        Task<bool> UpdateContentFeatureAsync(Guid contentId, string contentType, string featureCategory, string featureVector, double qualityScore);
         Task<List<MLContentFeature>> GetHighQualityContentAsync(string contentType, float minQualityScore, int limit = 100);
         Task<List<MLContentFeature>> GetViralContentAsync(float minViralScore, int limit = 50);
     }
