+++ Camply.Domain/Analytics/UserInterestDocument.cs
@@ -9,7 +9,7 @@ namespace Camply.Domain.Analytics
     {
         public ObjectId Id { get; set; }
         public Guid UserId { get; set; }
-        public Dictionary<string, float> Interests { get; set; } = new(); // Kategori -> Skor
+        public Dictionary<string, double> Interests { get; set; } = new(); // Kategori -> Skor
         public DateTime UpdatedAt { get; set; }
         public string Version { get; set; }
         public List<InterestTrend> History { get; set; } = new(); // Zaman içindeki değişim
